my moods
- ruben

+++ yourlifeyourvoice/Classes/BYCDatabase.h
@@ -3,7 +3,7 @@
 
 @interface BYCDatabase : NSObject
 -(id)initWithDatabasePath:(NSString *)databasePath;
--(long long)saveWithType:(BYCMoodType)type notes:(NSString*)notes reasons:(NSArray*)reasons;
+-(int64_t)saveWithType:(BYCMoodType)type notes:(NSString*)notes reasons:(NSArray*)reasons;
 -(BYCEntry*)getEntryWithUid:(int64_t)uid;
 -(NSArray*)getEntryPage:(NSInteger)page;
 @end
